pyproject.toml
[project]
name = "pal4vst"
version = "1.0.0"
authors = [
  { name="Sebastian Hartwig", email="sebastian@hartwig-bw.de" },
]
description = "PAL4VST python module."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "openmim",
    "mmengine",
    "mmcv>=2.0.0",
    "timm==0.6.5",
    "scikit-image",
    "mmpretrain>=1.0.0rc8",
    "mmdeploy==1.3.0",
    "transformers",
    "diffusers",
    "torchvision",
    "gdown"
]

[project.urls]
Homepage = "https://github.com/kopetri/PAL4VST"
Issues = "https://github.com/kopetri/PAL4VST/issues"